Deleted 'node_modules' directory
I learned that it is considered best practice to delete the 'node_modules' directory as it may be very large, the modules are often being updated, and are readily available to install using node.js when a project is pulled from the GitHub repo. All the modules required to be installed can be found in the 'package.json' file with the 'dependencies' key.