Created a proper template to render for the '/blog' route
Created the file 'blog.hbs', which acts as the template to be rendered when a request is made to the '/blog' route. The '/blog' route now displays a single blog post, which is the most recently made post by default, in an organized and well formatted manner. It also has a side bar to filter blog posts by category and select other blog posts. This functionality is added through a querying system that utilizes request parameters and querying, along with the 'blog-service' module, to select specific post objects and render them as defined in the 'blog.hbs' file.

Signed-off-by: Muhammad Ahmed <mahmed.4zad@gmail.com>