Bug fix for '/blog' route
The '/blog' route was not properly rendering posts when posts were selected form side panel. This was caused by the fact that the method 'getElementByID' in the 'blog-service' module was returning an array of one object, rather than a single object. Thus, this method was changed to return a single object and other changes were made in the 'server.js' code and the 'blog.hbs' Handlebars template to accomodate this change.

Signed-off-by: Muhammad Ahmed <mahmed.4zad@gmail.com>