Re-coded the 'blog-service' module to accomodate database integration
Defined postgresql models for user posts and post categories. Also, re-coded all the methods inside the 'blog-service' module to now accomdate the integration of the postgresql database. The methods have not yet been properly tested and thus, proper testing of each method and the creation of new unit tests is due to come in a future commit.

Signed-off-by: Muhammad Ahmed <mahmed.4zad@gmail.com>