Modified server code to accomodate database integration
Modified the server code in the 'server.js' file to fix possible bugs which may occur due to the changes that were made to the 'blog-service' module, in order to accomodate the postgresql database integration with the blogging web app. Most notably, the 'then' block of each 'blog-service' method call has added a condition to only render data if the data is not empty and otherwise, send a message to inform the user that no data exists in the database.

Signed-off-by: Muhammad Ahmed <mahmed.4zad@gmail.com>