Created and updated various 'views'
Created and updated various files in the '/views' directory, to accomodate the additional routes and functionality added to the server code, in the 'server.js' file. These changes allow for existing posts and categories to be deleted from their repective pages, found at the '/posts' and '/categories' routes. The changes also allow for the addition of new categories into the database, at the '/categories/add' route.

Signed-off-by: Muhammad Ahmed <mahmed.4zad@gmail.com>