Updated the form and back-end process for creating/uploading/adding a post
This involved amking sure the list of categories that someone selects from when creating a post are actual categories that exist, retrieved from the database. Furthermore, if no categories exist, then the user will be told so and be allowed to create a post without selecting any category, which the business/back-end logic will process appropriately.

Signed-off-by: Muhammad Ahmed <mahmed.4zad@gmail.com>