Critical bug fix in 'blog-service' module
Due to the use of the node module 'sequelize' to perform CRUD operations on the database, the methods in the 'blog-service' module were resolving even when they had no data to return (when an empty array of 0 objects was being returned). This was fixed using conditions placed in the 'then' block of each 'sequelize' method call, to have each 'blog-service' method return a rejected promise with an appropriate message.

Signed-off-by: Muhammad Ahmed <mahmed.4zad@gmail.com>