Minor bug fixes and code cleanup
The server code to add a new category, in 'server.js', for the '/categories/add' route was calling the wrong 'blog-service' module's method. It has now been fixed. Furthermore, there were a hanful of minor bugs in the 'blog-service' module, which have also been fixed. The display for blog posts that don't have an image has also been improved in the 'blog.hbs' file/view.

Signed-off-by: Muhammad Ahmed <mahmed.4zad@gmail.com>